+++ Reqnroll/BoDi/ObjectContainerException.cs
@@ -1,19 +1,11 @@
 ﻿using System;
 using System.Linq;
-using System.Runtime.Serialization;
 
 namespace Reqnroll.BoDi;
 
-[Serializable]
 public class ObjectContainerException : Exception
 {
     public ObjectContainerException(string message, Type[] resolutionPath) : base(GetMessage(message, resolutionPath))
-    {
-    }
-
-    protected ObjectContainerException(
-        SerializationInfo info,
-        StreamingContext context) : base(info, context)
     {
     }