+++ Reqnroll/ErrorHandling/ReqnrollException.cs
@@ -1,11 +1,9 @@
 using System;
-using System.Runtime.Serialization;
 
 // the exceptions are part of the public API, keep them in Reqnroll namespace
 // ReSharper disable once CheckNamespace
 namespace Reqnroll
 {
-    [Serializable]
     public class ReqnrollException : Exception
     {
         public ReqnrollException()
@@ -17,12 +15,6 @@ namespace Reqnroll
         }
 
         public ReqnrollException(string message, Exception inner) : base(message, inner)
-        {
-        }
-
-        protected ReqnrollException(
-            SerializationInfo info,
-            StreamingContext context) : base(info, context)
         {
         }
     }